Add check for cmr url validity, move function
