+++ pyQuARC/code/constants.py
@@ -1,5 +1,3 @@
-import os
-
 from colorama import Fore, Style
 from pathlib import Path
 
@@ -75,9 +73,3 @@ GCMD_LINKS = {
 }
 
 CMR_URL = "https://cmr.earthdata.nasa.gov"
-
-def get_cmr_url():
-    cmr_url = os.environ.get("CMR_URL", CMR_URL)
-    if not cmr_url.startswith("http"):
-        cmr_url = f"https://{cmr_url}"
-    return cmr_url
